debezium/dbz#1502 reinstate integration tests
