+++ journal-parsing/pom.xml
@@ -33,6 +33,11 @@
             <artifactId>mockito-junit-jupiter</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>ch.qos.logback</groupId>
+            <artifactId>logback-classic</artifactId>
+            <scope>test</scope>
+        </dependency>        
     </dependencies>
     <build>
         <resources>
@@ -57,4 +62,5 @@
             </testResource>
         </testResources>
     </build>
+
 </project>
