--- journal-parsing/src/it/java/com/fnz/db2/journal/retrieve/TriFunction.java
@@ -1,15 +1,0 @@
-package com.fnz.db2.journal.retrieve;
-
-import java.util.Objects;
-import java.util.function.Function;
-
-@FunctionalInterface
-public interface TriFunction<T, U, V, R> {
-
-	R apply(T t, U u, V v);
-
-	default <K> TriFunction<T, U, V, K> andThen(Function<? super R, ? extends K> after) {
-		Objects.requireNonNull(after);
-		return (T t, U u, V v) -> after.apply(apply(t, u, v));
-	}
-}
